pizza/requirements.txt
Django
squareup